format(repo): clean up
