+++ packages/db/schema/providers/gmail/provider.ts
@@ -7,7 +7,7 @@ import {
   varchar,
 } from "drizzle-orm/pg-core";
 
-import { emailProviderDetail } from "./email-provider-detail";
+import { emailProviderDetail } from "../../email-provider-detail";
 
 // token length: https://developers.google.com/identity/protocols/oauth2#:~:text=to%20service%20accounts.-,Token%20size,Refresh%20tokens%3A%20512%20bytes
 export const gmailProvider = pgTable("gmail_provider", {
