--- packages/db/schema/users.ts
@@ -1,6 +1,0 @@
-import { pgTable, serial, varchar } from "drizzle-orm/pg-core";
-
-export const users = pgTable("users", {
-  id: serial("id").primaryKey(),
-  email: varchar("email", { length: 255 }).notNull(),
-});
